requirements.txt
astroid==1.4.8
autopep8==1.2.4
Django==1.10.3
django-google-charts==0.1.1
django-rest==0.0.1
djangorestframework==3.5.3
gviz-data-table==1.0.2
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.5.2
pep8==1.7.0
pylint==1.6.4
pylint-django==0.7.2
pylint-plugin-utils==0.2.4
requests==2.12.4
six==1.10.0
wrapt==1.10.8
